Removed setup job. (#29)
Removed unnecessary job taking time from running the CI as fast as possible. The initial goal was to be able to setup the agent's environment and share it with other agents/VMs but that was my misunderstanding of the time and doing it not the right way. Since the other jobs were also setting up their BEAM environments, and working great on their own, I removed the setup job.